[libc] Split string and system modules
