+++ libc/system/cleanup.c
@@ -1,0 +1,1 @@
+void (* _cleanup)() = 0;  // called by exit()
