+++ libc/system/environ.c
@@ -1,0 +1,1 @@
+char ** environ;
