+++ libc/system/errno.c
@@ -1,0 +1,3 @@
+#include <errno.h>
+
+int errno;
