+++ libc/system/getegid.c
@@ -1,0 +1,9 @@
+#include <unistd.h>
+
+int
+getegid(void)
+{
+   int egid;
+   _getgid(&egid);
+   return egid;
+}
