+++ libc/system/geteuid.c
@@ -1,0 +1,9 @@
+#include <unistd.h>
+
+int
+geteuid(void)
+{
+   int euid;
+   _getuid(&euid);
+   return euid;
+}
