+++ libc/system/getgid.c
@@ -1,0 +1,7 @@
+#ifdef L_getgid
+int getgid(void)
+{
+   int egid;
+   return __getgid(&egid);
+}
+#endif
