+++ libc/system/getpgid.c
@@ -1,0 +1,9 @@
+#ifdef L_getpgrp
+#include <unistd.h>
+
+int
+getpgrp(void)
+{
+   return getpgid(0);
+}
+#endif
