+++ libc/system/getppid.c
@@ -1,0 +1,11 @@
+#ifdef L_getppid
+#include <unistd.h>
+
+int
+getppid(void)
+{
+   int ppid;
+   __getpid(&ppid);
+   return ppid;
+}
+#endif
