+++ libc/system/getuid.c
@@ -1,0 +1,8 @@
+#include <unistd.h>
+
+uid_t
+getuid(void)
+{
+   int euid;
+   return _getuid(&euid);
+}
