+++ libc/system/mkfifo.c
@@ -1,0 +1,9 @@
+#ifdef L_mkfifo
+#include <sys/stat.h>
+
+int
+mkfifo(char * path, int mode)
+{
+   return mknod(path, mode | S_IFIFO, 0);
+}
+#endif
