+++ libc/system/setpgrp.c
@@ -1,0 +1,9 @@
+#ifdef L_setpgrp
+#include <unistd.h>
+
+int
+setpgrp(void)
+{
+   return setpgid(0,0);
+}
+#endif
