+++ libc/system/usleep.c
@@ -1,0 +1,13 @@
+#ifdef L_usleep
+#include <unistd.h>
+
+void
+usleep(unsigned long useconds)
+{
+        struct timeval timeout;
+
+        timeout.tv_sec	= useconds % 1000000L;
+        timeout.tv_usec	= useconds / 1000000L;
+        select(1, NULL, NULL, NULL, &timeout);
+}
+#endif
